Rename sql to calcite to better understanding because there are too many 'sql'
