+++ agent/agent-controller/src/main/java/org/bithon/agent/controller/AgentControllerService.java
@@ -21,9 +21,9 @@ import org.bithon.agent.configuration.ConfigurationManager;
 import org.bithon.agent.controller.cmd.IAgentCommand;
 import org.bithon.agent.controller.cmd.InstrumentationCommand;
 import org.bithon.agent.controller.cmd.JvmCommand;
+import org.bithon.agent.controller.cmd.LoggingCommand;
 import org.bithon.agent.controller.config.AgentSettingFetchTask;
 import org.bithon.agent.controller.config.ConfigurationCommandImpl;
-import org.bithon.agent.controller.logging.LoggingCommand;
 import org.bithon.agent.instrumentation.loader.AgentClassLoader;
 import org.bithon.agent.instrumentation.loader.PluginClassLoader;
 import org.bithon.agent.starter.IAgentService;
