+++ server/web-service/src/main/java/org/bithon/server/web/service/agent/api/AgentDiagnosisApi.java
@@ -54,12 +54,12 @@ import org.bithon.server.web.service.WebServiceModuleEnabler;
 import org.bithon.server.web.service.agent.sql.AgentSchema;
 import org.bithon.server.web.service.agent.sql.table.AgentServiceProxyFactory;
 import org.bithon.server.web.service.agent.sql.table.IPushdownPredicateProvider;
+import org.bithon.server.web.service.common.calcilte.SqlExecutionContext;
+import org.bithon.server.web.service.common.calcilte.SqlExecutionEngine;
+import org.bithon.server.web.service.common.calcilte.SqlExecutionResult;
 import org.bithon.server.web.service.common.output.IOutputFormatter;
 import org.bithon.server.web.service.common.output.JsonCompactOutputFormatter;
 import org.bithon.server.web.service.common.output.TabSeparatedOutputFormatter;
-import org.bithon.server.web.service.common.sql.SqlExecutionContext;
-import org.bithon.server.web.service.common.sql.SqlExecutionEngine;
-import org.bithon.server.web.service.common.sql.SqlExecutionResult;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Conditional;
 import org.springframework.http.HttpStatus;
