+++ server/web-service/src/main/java/org/bithon/server/web/service/common/calcilte/IUpdatableTable.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.web.service.common.sql;
+package org.bithon.server.web.service.common.calcilte;
 
 import org.bithon.component.commons.expression.IExpression;
